some refactors
